+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- fix: Add Trace Headers below iOS 14.0 (#1302)
+
 ## 7.2.5
 
 - fix: Swizzling crash on iOS 13 (#1297)